Move the absorption probs method to base estimator
